Extract app error presenter
